feat: verify a returned eSign signature, with OCSP/CRL revocation
Adds eSign.verifyEsignResponseHash(sha256Hex, esignRespXml, checkRevocation),
which checks the PKCS#7 the gateway returned against the hash that was sent,
reports the signer certificate, and optionally asks the CA whether that
certificate is revoked. It reads nothing from disk and never touches the PDF,
so it covers hash-mode signing over JSON or any other payload as well as PDF
signing.

- eSignVerifier: CMS verification, issuer resolution (from the PKCS#7 or
  downloaded via the AIA extension), OCSP with a CRL fallback. Never throws;
  setHttpTimeouts is the only public knob (15 s / 45 s defaults).
- eSignVerificationResult: isValid() plus the signature, certificate,
  revocation and gateway-response fields. Certificate validity is evaluated at
  the CMS signing time when present, because eSign OTP certificates live about
  30 minutes and are already expired by the time anyone verifies.
- eSignResponseInfo: <EsignResp> parser and core validation of the gateway's
  enveloped XML-DSig. That XML result is advisory and excluded from isValid(),
  since re-serialising the response is enough to break it.
- The second argument accepts either the response XML or a bare Base64/PEM
  PKCS#7. With several DocSignature elements, the one whose signed digest
  matches sha256Hex is verified, so no document id is needed.

Also in this merge:

- Signature layer2 text is word-wrapped before the font is reduced, so one long
  line no longer shrinks the whole block to an unreadable size, and lines are no
  longer silently dropped. In legacy-tick mode the top 30% of the box is
  reserved for the viewer's validity header instead of being overprinted.
- Page-level coordinates parse as float. They are PDF points and may be
  fractional (613.91); the renderer already read them with Float.valueOf, so
  integer parsing rejected coordinates the renderer would have accepted.
- A supplied pdfViewerLicence is handed to ValidateKitLicence by reflection,
  restoring backward compatibility with older emPDFViewer kits. The encryption
  key stays unconditional. ValidateKitLicence ships as a .class only and is
  excluded by .gitignore, so the reflective call logs a warning and continues
  when it is absent.

Documentation for all of the above in README.md, QUICK_START.md (new "Verifying
a Signature" section plus the API-reference entry) and FRAMEWORK_INTEGRATION.md
(a framework-agnostic callback example).

+++ src/com/emudhra/esign/eSignUtility.java
@@ -416,10 +416,13 @@ public final class eSignUtility {
             int pageNumber = Integer.parseInt(signature[0]);
             String position = signature[1];
             String[] rect = position.split(",");
-            int x1 = Integer.parseInt(rect[0]);
-            int y1 = Integer.parseInt(rect[1]);
-            int x2 = Integer.parseInt(rect[2]);
-            int y2 = Integer.parseInt(rect[3]);
+            // Coordinates are PDF points and may be fractional (e.g. 613.91);
+            // the renderer already reads them with Float.valueOf. Page number
+            // stays an int. The page-bounds check below is unchanged.
+            float x1 = Float.parseFloat(rect[0]);
+            float y1 = Float.parseFloat(rect[1]);
+            float x2 = Float.parseFloat(rect[2]);
+            float y2 = Float.parseFloat(rect[3]);
             Rectangle r = reader.getPageSizeWithRotation(pageNumber);
             if (isContentSearch) {
                 if (!(r.getWidth() >= x1 && r.getWidth() >= x2 && r.getHeight() >= y1 && r.getHeight() >= y2)) {
